casi terminado, arrelar delete calif

+++ backend/main/resources/Calificacion.py
@@ -1,7 +1,9 @@
 from flask_restful import Resource
 from flask import request, jsonify
-from main.models import CalificacionModel
 from .. import db
+from main.models import CalificacionModel
+from main.models import UsuarioModel
+from flask_jwt_extended import verify_jwt_in_request, get_jwt
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from main.auth.decorators import admin_required
 
@@ -17,13 +19,21 @@ class Calificacion(Resource):
         calificacion = db.session.query(CalificacionModel).get_or_404(id)
         return calificacion.to_json()
     
-    @jwt_required
-    @admin_required
+    @jwt_required()
     def delete(self,id):
         calificacion = db.session.query(CalificacionModel).get_or_404(id)
-        db.session.delete(calificacion)
-        db.session.commit()
-        return '',204
+        identidad_usuario = get_jwt_identity()
+        calificacion.calificacionId = identidad_usuario
+        claims = get_jwt()
+        if calificacion.usuario_id == identidad_usuario or claims['rol'] == "admin":
+            try:
+                db.session.delete(calificacion)
+                db.session.commit()
+            except Exception as error:
+                return 'Formato Invalido', 204
+            return calificacion.to_json(), 201
+        else:
+            return 'No tienes permisos para realizar esta accion.', 403
     
     @jwt_required
     def put(self,id):
@@ -40,9 +50,19 @@ class Calificaciones(Resource):
         calificaciones = db.session.query(CalificacionModel).all()
         return jsonify ([calificacion.to_json_short() for calificacion in calificaciones])
 
-    @jwt_required
+    @jwt_required()
     def post(self):
         calificacion = CalificacionModel.from_json(request.get_json())
-        db.session.add(calificacion)
-        db.session.commit()
-        return calificacion.to_json(), 201
+        identidad_usuario = get_jwt_identity
+        calificacion.calificacionId = identidad_usuario
+        claims = get_jwt()
+
+        if identidad_usuario and claims["rol"] != "admin":
+            try:
+                db.session.add(calificacion)
+                db.session.commit()
+            except Exception as error:
+                return 'Formato incorrecto', 400
+            return calificacion.to_json(), 201
+        else:
+            return 'No dispones de permisos para realizar esta accion.!', 403
